Moves out of JSON.parse try/catch block, user's function handler
Whenever a user has any error that throws an exception in her function handler,
that is not caught up in her code, it gets caught by swarm code
and hides the issue from the user's eyes.

This fix takes the invocation of the user's callbacks out try/catch
block for JSON.parse, so that any error is shown up immediately
in the browser's console of the user.